Struct memory_manager::block::BlockDescriptor

pub struct BlockDescriptor<'a> {
    pub start: *mut u8,
    pub free: *mut u8,
    // some fields omitted
}

Memory block: collection of objects.

Block Layout

┏━━━━━━━━━━┯━━━━━━━━━━┯━━━━━┯━━━━━━━━━━┯━━━━━━━━━━━━━━┓
┃ reserved │ object 0 │ ... │ object N │ not used yet ┃
┗━━━━━━━━━━┷━━━━━━━━━━┷━━━━━┷━━━━━━━━━━┷━━━━━━━━━━━━━━┛
           ↑                           ↑
           │                           │
         start                        free

Reserved space: we may reserved some space to avoid malloc, etc.

object 1 ~ N: objects managed by this memory-manager.

not used yet: for future allocation, or wasted due to fragmentation.

Fields
start: *mut u8

The starting address for this block.

Invariant: unless start == free, at start there is a valid ObjectDescriptor.

free: *mut u8

The first free address in this block.

Invariant: no pointers in the same block is after free.

Implementations

impl<'a> BlockDescriptor<'a>

pub const SIZE: usize

Size of a Block.

pub const SIZE_IN_WORDS: usize

Size of a Block in Words (usizes).

pub fn new(start: *mut u8) -> Self

Constructor for BlockDescriptor.

pub fn objects(&self) -> ObjectIterator<'a>

Iterate on the objects in this block.