Enum memory_manager::primitives::MMapError

pub enum MMapError {
    InvalidArguments,
    TryAgain,
    NoMemory,
    LengthOverflow,
    UnknownError(u32),
    NoError,
}

Common errors from mmap.

Error codes on Windows is far more complicated then errno on UNIX-like systems. Also, we don't know what error code can be caused by an API call, for they are both not documented and officially-announced as unstable. Therefore on Windows, we try our best to match on the error, but likely some errors would still end up as an UnknownError.

Note: items marked as UNIX-specific will not show up on Windows.

Variants

InvalidArguments

Arguments provided to mmap is invalid.

TryAgain

(UNIX-specific) Too much memory has been locked.

NoMemory

(UNIX-specific) Any of the following applies:

No memory available, or

Maximum number of mappings exceeded, or

RLIMIT_DATA exceeded

LengthOverflow

Number of pages overflows unsigned long. (32-bit platform only, UNIX-specific)

UnknownError(u32)

Errors not recognized, with the raw error code on the host system.

NoError

No error at all, NOT EXPECTED. Whenever received, this should be considered as a bug in the implementation.

Implementations

impl MMapError

pub fn from_errno(e: c_int) -> MMapError

Get MMapError from an errno value.

pub unsafe fn get() -> MMapError

Get MMapError from the current errno value.